// acc-sync-app/src/api/models/SignupDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * kartografia-backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SignupDTO
 */
export interface SignupDTO {
    /**
     * 
     * @type {string}
     * @memberof SignupDTO
     */
    firstName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SignupDTO
     */
    lastName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SignupDTO
     */
    username?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SignupDTO
     */
    email?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SignupDTO
     */
    password?: string | null;
}

/**
 * Check if a given object implements the SignupDTO interface.
 */
export function instanceOfSignupDTO(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function SignupDTOFromJSON(json: any): SignupDTO {
    return SignupDTOFromJSONTyped(json, false);
}

export function SignupDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): SignupDTO {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'firstName': !exists(json, 'firstName') ? undefined : json['firstName'],
        'lastName': !exists(json, 'lastName') ? undefined : json['lastName'],
        'username': !exists(json, 'username') ? undefined : json['username'],
        'email': !exists(json, 'email') ? undefined : json['email'],
        'password': !exists(json, 'password') ? undefined : json['password'],
    };
}

export function SignupDTOToJSON(value?: SignupDTO | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'firstName': value.firstName,
        'lastName': value.lastName,
        'username': value.username,
        'email': value.email,
        'password': value.password,
    };
}